Convert Markdown to .tsx and adjust mobile styling
