+++ components/fcc/markdown/Markdown.module.scss
@@ -1,6 +1,24 @@
 @use '../../../styles/globalStyles.module.scss' as global;
 
+$borderRadius: 20px;
+
 .container {
+	width: 100%;
+	height: 100%;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	gap: 2%;
+}
+
+@media (min-aspect-ratio: 1) {
+	.container {
+		flex-direction: row;
+	}
+}
+
+.editDialogue {
 	width: 90%;
 	height: 90%;
 	display: flex;
@@ -11,7 +29,8 @@
 }
 
 @media (min-aspect-ratio: 1) {
-	.container {
+	.editDialogue {
+		height: 100%;
 		flex-direction: row;
 	}
 }
@@ -20,24 +39,23 @@
 	position: relative;
 	display: flex;
 	flex-direction: column;
+	justify-content: flex-start;
 	align-items: center;
-	justify-content: space-between;
-	border-radius: 10px;
+	border-radius: $borderRadius;
 	transition: width var(--medium-transition), opacity var(--medium-transition), height var(--medium-transition);
 	transition-timing-function: ease-in-out;
 }
 
 .header {
+	position: relative;
 	display: flex;
-	flex-direction: row;
 	align-items: center;
-	flex-wrap: nowrap;
-	gap: 2vmin;
+	gap: 1em;
 	width: 100%;
-	height: 5em;
-	padding: 0 2vmin;
-	border-top-left-radius: 10px;
-	border-top-right-radius: 10px;
+	min-height: 8svh;
+	padding: 1vmin;
+	border-top-left-radius: $borderRadius;
+	border-top-right-radius: $borderRadius;
 	background-color: #142939;
 	overflow-y: hidden;
 	overflow-x: auto;
@@ -50,33 +68,44 @@
 	font-weight: bold;
 }
 
-.editor {
-	background-color: #D7DFEA;
-}
-
 .editor .header {
 	justify-content: flex-end;
-}
-
-.preview {
-	background-color: #F7C9C9;
+	flex-direction: row;
 }
 
 .preview .header {
-	justify-content: flex-end;
+	flex-direction: row-reverse;
+	justify-content: flex-start;
 }
 
 .paneChild {
 	display: block;
-	width: calc(100% - 2em);
-	height: calc(100% - 2em);
+	width: 100%;
+	height: 100%;
 	padding: 1em;
-	margin: 1em;
 	overflow-y: auto;
 	resize: none;
-	border-bottom-left-radius: 10px;
+	border-bottom-left-radius: $borderRadius;
 	scrollbar-width: thin;
 	color: global.$Black;
+}
+
+.contentContainer {
+	position: relative;
+	height: 100%;
+	width: 100%;
+	border-bottom-left-radius: $borderRadius;
+	border-bottom-right-radius: $borderRadius;
+	padding: 2svh;
+	overflow-y: hidden;
+}
+
+.editor .contentContainer {
+	background-color: #D7DFEA;
+}
+
+.preview .contentContainer {
+	background-color: #F7C9C9;
 }
 
 .textarea {
@@ -92,17 +121,8 @@
 
 @media (min-aspect-ratio: 1) {
 	.preview .header {
+		flex-direction: row;
 		justify-content: flex-start;
-	}
-}
-
-@media (min-aspect-ratio: 1) {
-	.textarea {
-		height: 85%;
-	}
-
-	.output {
-		height: 85%;
 	}
 }
 
@@ -190,16 +210,25 @@
 
 .bothPanes {
 	width: 90%;
-	height: 50%;
+	height: 49%;
 	opacity: 1;
 }
 
+.toggled .hidden {
+	width: 90%;
+	opacity: 1;
+}
 
-.toggled,
+.toggled {
+	width: 90%;
+	height: 90%
+}
+
 .hidden {
 	width: 90%;
-	height: 50%;
+	height: 0;
 	opacity: 1;
+	overflow-y: hidden;
 }
 
 @media (min-aspect-ratio: 1) {
@@ -224,18 +253,19 @@
 }
 
 .paneToggler {
-	display: none;
+	display: block;
 }
 
 .paneToggler svg {
 	display: block;
 	margin: auto;
 	font-size: 2rem;
+	transform: rotate(90deg)
 }
 
 @media (min-aspect-ratio: 1) {
-	.paneToggler {
-		display: block;
+	.paneToggler svg {
+		transform: none;
 	}
 }
 
@@ -280,7 +310,7 @@
 @media (min-aspect-ratio: 1) {
 	.updater {
 		svg {
-			rotate: 0deg;
+			transform: none;
 		}
 	}
 }
@@ -288,10 +318,10 @@
 .dlContainer {
 	backdrop-filter: brightness(0.15) blur(20px);
 	position: absolute;
-	height: 100%;
+	height: 100svh;
 	width: 100%;
 	top: 0;
-	border-radius: 10px;
+	border-radius: $borderRadius;
 	animation-name: dlContainerFadeIn;
 	animation-duration: 0.2s;
 	animation-timing-function: ease-in-out;
@@ -315,7 +345,7 @@
 	width: 2em;
 	height: 2em;
 	margin: 1em;
-	border-radius: 10px;
+	border-radius: $borderRadius;
 	color: global.$White;
 	background-color: global.$Black;
 	border: 2px solid global.$White;
@@ -357,11 +387,11 @@
 	width: 100%;
 	max-width: 500px;
 	padding: 0.2em;
-	border-radius: 10px;
+	border-radius: $borderRadius;
 	background-color: global.$Black;
 	border: 2px solid global.$White;
 	overflow-y: auto;
-	gap: 10px;
+	gap: $borderRadius;
 	scrollbar-width: thin;
 }
 
@@ -381,8 +411,8 @@
 	padding: 0.8em;
 	font-size: inherit;
 	background-color: global.$slightlyLighter;
-	border-radius: 10px;
-	gap: 10px;
+	border-radius: $borderRadius;
+	gap: $borderRadius;
 
 	svg {
 		display: block;
